refactor: more better structure for verification module
